graph folder restructured
