+++ .gitignore
@@ -216,3 +216,7 @@ questdb/
 node_modules/
 package-lock.json
 package.json
+
+backend/graph/README.md
+backend/graph/test.py
+backend/graph/graph_diagram.png
