--- backend/graph/nodes/detect_language.py
@@ -1,16 +1,0 @@
-from langchain_core.messages import HumanMessage
-from state import State
-from src.utils.nlp import detect_language
-
-
-def detect_language_node(state: State):
-    """Detect the language of the last user message."""
-    for msg in reversed(state.messages):
-        if isinstance(msg, HumanMessage) and isinstance(msg.content, str):
-            try:
-                detected = detect_language(msg.content)
-            except RuntimeError:
-                return {"detected_lang": "es"}
-            return {"detected_lang": detected}
-
-    return {"detected_lang": "es"}
