--- backend/graph/tools/test_tool.py
@@ -1,8 +1,0 @@
-from langchain.tools import tool
-
-
-@tool
-def test_tool():
-    """Performs a test whenever the user wants to perform a test"""
-
-    return "Test succeeded!"
